refactor(@142vip/fairy-cli): 模块结构调整，简化导出代码和cli处理流程
- 简化build配置，编译、导出更少逻辑代码
- 更新README文档
