+++ packages/fairy-cli/CHANGELOG.md
@@ -1,4 +1,7 @@
-# Change Log
+# Changelog
+
+All notable changes to this project will be documented in this file. See [commit-and-tag-version](https://github.com/absolute-version/commit-and-tag-version) for commit guidelines.
+
 
 ## v0.0.3-alpha.2 (2024-08-08)
