+++ packages/fairy-cli/build.config.ts
@@ -1,9 +1,9 @@
 import { defineBuildConfig } from 'unbuild'
 
+// 参考：https://github.com/unjs/unbuild
 export default defineBuildConfig({
   entries: [
     'src/index',
-    'src/fairy-cli',
   ],
   declaration: true,
   clean: true,
