+++ packages/fairy-cli/cli.mjs
@@ -1,3 +1,4 @@
-#!/usr/bin/env node
+// eslint-disable-next-line antfu/no-import-dist
+import { fairyCliMain } from './dist/index.mjs'
 
-import './dist/fairy-cli.mjs'
+fairyCliMain()
