+++ packages/fairy-cli/package.json
@@ -20,8 +20,7 @@
   "module": "./dist/index.mjs",
   "types": "./dist/index.d.ts",
   "bin": {
-    "fa": "./cli.mjs",
-    "vae": "./cli.mjs"
+    "fa": "./cli.mjs"
   },
   "files": [
     "*.mjs",
