+++ packages/fairy-cli/src/commands/index.ts
@@ -7,3 +7,4 @@ export * from './publish'
 export * from './release'
 export * from './turbopack'
 export * from './install'
+export * from './sync'
