+++ packages/fairy-cli/src/commands/install.ts
@@ -1,6 +1,7 @@
-import { execChildProcess } from '../utils'
+import type { Command } from 'commander'
+import { CliCommandEnum, execChildProcess } from '../shared'
 
-export interface InstallOptions {
+interface InstallOptions {
   pnpm: boolean
   npm: boolean
   registry?: string
@@ -12,7 +13,7 @@ export interface InstallOptions {
  * - npm
  * - pnpm
  */
-export function execInstall(args: InstallOptions) {
+function execInstall(args: InstallOptions) {
   console.log(111, args)
   // pnpm i --frozen-lockfile --registry https://registry.npmmirror.com
   // npm ci
@@ -24,4 +25,22 @@ export function execInstall(args: InstallOptions) {
     // pnpm下载
     execChildProcess(`pnpm i ${args.update ? '' : '--frozen-lockfile'} --registry ${args.registry}`)
   }
+}
+
+/**
+ * install 命令入口
+ * @param program
+ */
+export async function installMain(program: Command) {
+  program
+    .command(CliCommandEnum.INSTALL)
+    .aliases(['i', 'add'])
+    .description('pnpm ci dependencies')
+    .option('--pnpm', 'use pnpm ', true)
+    .option('--npm', 'use npm', false)
+    .option('--update', 'update lockfile，not use --frozen-lockfile', false)
+    .option('--registry', 'pnpm registry address，default ali cdn', 'https://registry.npmmirror.com')
+    .action((args: InstallOptions) => {
+      execInstall(args)
+    })
 }
