+++ packages/fairy-cli/src/commands/lint.ts
@@ -1,11 +1,13 @@
-import { execCommand } from '../utils'
+import { execShell } from '@142vip/common'
+import type { Command } from 'commander'
+import { CliCommandEnum } from '../shared'
 
-export interface LintOptions {
+interface LintOptions {
   markdown: boolean
   fix: boolean
 }
 
-export async function execLink(args: LintOptions) {
+async function execLink(args: LintOptions) {
   // auto fix
   if (args.fix) {
     console.log('自动修复')
@@ -20,13 +22,29 @@ export async function execLink(args: LintOptions) {
 /**
  * 代码格式化
  */
-export async function doCodeLint() {
-  await execCommand('npx eslint . --fix')
+async function doCodeLint() {
+  await execShell('npx eslint . --fix')
 }
 
 /**
  * markdown文档格式化
  */
-export function doMdLink() {
+function doMdLink() {
   // markdownlint
+}
+
+/**
+ *
+ * @param program
+ */
+export async function lintMain(program: Command) {
+  program
+    .command(CliCommandEnum.LINT)
+    .description('根据Eslint检查代码风格，支持代码格式化')
+    .option('-m,--markdown', '格式化markdown文件')
+    .option('-c,--config', 'Eslint配置文件路径', false)
+    .option('-f --fix', '支持自动修复', false)
+    .action(async (args: LintOptions) => {
+      await execLink(args)
+    })
 }
