+++ packages/fairy-cli/src/commands/login.ts
@@ -1,18 +1,20 @@
-import { execChildProcess } from '../utils'
+import process from 'node:process'
+import type { Command } from 'commander'
+import { CliCommandEnum, execChildProcess } from '../shared'
 
-export enum LoginPlatformEnum {
+enum LoginPlatformEnum {
   DOCKER = 'docker',
   NPM = 'npm',
 }
 
-export enum RegistryURLEnum {
+enum RegistryURLEnum {
   DOCKER = 'https://registry.cn-hangzhou.aliyuncs.com',
   NPM = 'https://registry.npmjs.org',
   VIP_DOCKER = 'registry.cn-hangzhou.aliyuncs.com',
   VIP_NPM = 'https://registry.142vip.com',
 }
 
-export interface LoginOptions {
+interface LoginOptions {
   userName?: string
   password?: string
   registryUrl?: string
@@ -20,13 +22,13 @@ export interface LoginOptions {
   vip?: string
 }
 
-export interface DockerOptions extends LoginOptions {
+interface DockerOptions extends LoginOptions {
 }
 
-export interface NpmOptions extends Omit<LoginOptions, 'userName' | 'password'> {
+interface NpmOptions extends Omit<LoginOptions, 'userName' | 'password'> {
 }
 
-export function execLogin(platform: LoginPlatformEnum, args: LoginOptions) {
+function execLogin(platform: LoginPlatformEnum, args: LoginOptions) {
   if (LoginPlatformEnum.DOCKER === platform) {
     loginDocker(args)
   }
@@ -66,4 +68,24 @@ function loginNpm(args: NpmOptions) {
 
   // npm login --registry  https://registry.npmjs.org
   execChildProcess(`npm login --registry ${registryUrl}`)
+}
+
+/**
+ * login命令入口
+ */
+export async function loginMain(program: Command) {
+  program
+    .command(`${CliCommandEnum.LOGIN} <platform>`)
+    .description('登录远程平台，支持Docker和Npm')
+    .option('-u,--username', '登录账号，docker登录时有效')
+    .option('-p,--password', '登录密码，docker登录时有效')
+    .option('--registry-url', 'registry address')
+    .option('--vip', '142vip专用业务账号', false)
+    .action((platform: LoginPlatformEnum, args: LoginOptions) => {
+      if (![LoginPlatformEnum.NPM, LoginPlatformEnum.DOCKER].includes(platform)) {
+        console.error('login命令只支持Docker和Npm平台，使用格式 login docker|npm')
+        process.exit(1)
+      }
+      execLogin(platform, args)
+    })
 }
