+++ packages/fairy-cli/src/commands/turbopack.ts
@@ -1,13 +1,15 @@
 import fsSync from 'node:fs'
 import * as process from 'node:process'
 import { execShell } from '@142vip/common'
+import type { Command } from 'commander'
+import { CliCommandEnum } from '../shared'
 
-export interface TurboPackOptions {
+interface TurboPackOptions {
   dev?: boolean
   build?: boolean
 }
 
-export async function execTurboPack(repoNames: string[], args: TurboPackOptions) {
+async function execTurboPack(repoNames: string[], args: TurboPackOptions) {
   // 判断当前目录下是否有turbo.json配置文件
   if (!fsSync.existsSync('turbo.json')) {
     console.log('项目根目录下缺少turbo.json配置文件，查看Turborepo官网：<https://turbo.build/repo/docs>')
@@ -33,4 +35,20 @@ export async function execTurboPack(repoNames: string[], args: TurboPackOptions)
  */
 function getFilterRepo(repoNames: string[]) {
   return repoNames.map(name => `--filter=${name}`).join(' ')
+}
+
+/**
+ * turbo命令入口
+ * @param program
+ */
+export async function turboPackMain(program: Command) {
+  program
+    .command(CliCommandEnum.TURBO)
+    .description('exec turbo pack command')
+    .argument('[repoName...]', '需要使用Turbo管理的项目名，支持多个')
+    .option('--dev', '执行dev命令', false)
+    .option('--build', '执行build命令', false)
+    .action(async (repoNames: string[], args: TurboPackOptions) => {
+      await execTurboPack(repoNames, args)
+    })
 }
