+++ packages/fairy-cli/src/shared/cli-command.enum.ts
@@ -1,0 +1,12 @@
+export enum CliCommandEnum {
+  LOGIN = 'login',
+  RELEASE = 'release',
+  CHANGELOG = 'changelog',
+  PUBLISH = 'publish',
+  CLEAN = 'clean',
+  LINT = 'lint',
+  DEPLOY = 'deploy',
+  TURBO = 'turbo',
+  INSTALL = 'install',
+  SYNC = 'sync',
+}
