+++ packages/fairy-cli/src/shared/exec-command.ts
