+++ packages/fairy-cli/src/shared/index.ts
@@ -1,2 +1,3 @@
+export * from './cli-command.enum'
 export * from './exec-command'
 export * from './monorepo'
